Extract DataPageBuilder to test common
